+++ src/Entities/ApiKey.php
@@ -19,20 +19,6 @@ use Rebilly\Rest\Entity;
  */
 final class ApiKey extends Entity
 {
-    public const DATETIME_FORMAT_MYSQL = 'mysql';
-
-    public const DATETIME_FORMAT_ISO8601 = 'iso8601';
-
-    public const MSG_UNEXPECTED_DATETIME_FORMAT = 'Unexpected datetime format. Only %s formats support';
-
-    /**
-     * @return array
-     */
-    public static function datetimeFormats()
-    {
-        return [self::DATETIME_FORMAT_MYSQL, self::DATETIME_FORMAT_ISO8601];
-    }
-
     /**
      * @return string;
      */
@@ -49,30 +35,6 @@ final class ApiKey extends Entity
     public function setDescription($value)
     {
         return $this->setAttribute('description', $value);
-    }
-
-    /**
-     * @return string;
-     */
-    public function getDatetimeFormat()
-    {
-        return $this->getAttribute('datetimeFormat');
-    }
-
-    /**
-     * @param string $value
-     *
-     * @return $this
-     */
-    public function setDatetimeFormat($value)
-    {
-        if (!in_array($value, self::datetimeFormats(), true)) {
-            throw new DomainException(
-                sprintf(self::MSG_UNEXPECTED_DATETIME_FORMAT, implode(', ', self::datetimeFormats()))
-            );
-        }
-
-        return $this->setAttribute('datetimeFormat', $value);
     }
 
     /**