+++ tests/TestCase.php
@@ -668,7 +668,7 @@ abstract class TestCase extends Framework\TestCase
             case 'browserData':
                 return Entities\BrowserData::createFromData([
                     'colorDepth' => self::randomElements([1, 4, 8, 15, 16, 24, 32, 48])[0],
-                    'javaEnabled' => true,
+                    'isJavaEnabled' => true,
                     'language' => 'en-US',
                     'screenHeight' => random_int(0, 999999),
                     'screenWidth' => random_int(0, 999999),